added firebase.js file
