+++ src/components/userprofile/Login.vue
@@ -53,7 +53,7 @@
 </template>
 
 <script>
-import firebase from "firebase";
+import firebase from 'firebase';
 
 export default {
   data() {
